Insert whitespace before help-inline span
Rails will add error messages that are meant to be read in the context
of the label. Without some whitespace between the tags, page content
will be joined and automated tests will not find expected text.

i.e. When running a capybara spec, where a name field that cannot be
blank, the errors would result in a page text of

"Namecan't be blank"

Adding whitespace before the span tag should not result in a visible
difference in the browser.